make action spaces and factories public by putting them in the env/sim class
