+++ ai/environments/_simulator_wrapper.py
@@ -9,6 +9,14 @@ import ai.environments as environments
 class SimulatorWrapper(environments.Base):
     """Environment that wraps a simulator, exposing it as an environment."""
 
+    class Factory(environments.Factory):
+        def __init__(self, factory: simulators.Factory):
+            self._factory = factory
+
+        def __call__(self) -> "SimulatorWrapper":
+            return SimulatorWrapper(self._factory())
+
+
     def __init__(self, simulator: simulators.Base):
         """
         Args:
@@ -44,14 +52,6 @@ class SimulatorWrapper(environments.Base):
         self._simulator.render(self._state)
 
     @classmethod
-    def get_factory(cls, factory: simulators.Factory) -> environments.Factory:
+    def get_factory(cls, factory: simulators.Factory) -> "SimulatorWrapper.Factory":
         """Wraps a simulator factory into an environment factory."""
-        return Factory(factory)
-
-
-class Factory(environments.Factory):
-    def __init__(self, factory: simulators.Factory):
-        self._factory = factory
-
-    def __call__(self) -> SimulatorWrapper:
-        return SimulatorWrapper(self._factory())
+        return SimulatorWrapper.Factory(factory)
