+++ ai/environments/action_spaces/__init__.py
@@ -2,7 +2,6 @@
 
 from ._base import Base
 from ._discrete import Discrete
-from ._discrete_gym_wrapper import DiscreteGymWrapper
 
 
-__all__ = ["Base", "Discrete", "DiscreteGymWrapper"]
+__all__ = ["Base", "Discrete"]
